+++ velesresearch/models.py
@@ -834,6 +834,7 @@ class SurveyModel(BaseModel):
         title (str | None): The title of the survey.
         tocLocation (str): The location of the table of contents. Can be 'left' (default), 'right'. See `showTOC`.
         triggers (str | None): Triggers for the survey. Usually not necessary. See <https://surveyjs.io/form-library/documentation/design-survey/conditional-logic#conditional-survey-logic-triggers>.
+        UrlParameters (list[str] | None): The URL parameters to be expected and saved. Default is None.
         validateVisitedEmptyFields (bool): Whether to validate empty fields that had been clicked, and unclicked empty. Default is False.
         width (str | None): Width of the survey in CSS units. Default is None (inherit from the container).
         widthMode (str): The mode of the width. Can be 'auto' (default; the width is set by the content), 'static', 'responsive'.
@@ -911,6 +912,7 @@ class SurveyModel(BaseModel):
     title: str | None = None
     tocLocation: str = "left"
     triggers: list[dict] | None = None
+    UrlParameters: list[str] | None = None
     validateVisitedEmptyFields: bool = False
     width: str | None = None
     widthMode: str = "auto"