+++ NEWS.md
@@ -3,6 +3,7 @@
 ## future release
 
 * Hot reloading now only reloads the survey, not the whole page.
+* Added `UrlParameters` argument to the `Survey` class to specify what parameters should be saved in the database.
 
 ## 0.3.2